requirements.txt
python-telegram-bot>=20.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
APScheduler>=3.10.0
